+++ src/Hooks/useEmptyFields.ts
@@ -37,6 +37,7 @@ interface IuseEmptyFields {
 
 interface useEmptyFields_Return {
     isMising: boolean
+    isMissing: boolean
     MissingFields: Array<string>
     Length: number
 }
@@ -55,6 +56,7 @@ function useEmptyFields({
     if (inputObject === undefined || inputObject === null) {
         return {
             isMising: true,
+            isMissing: true,
             MissingFields: ReqiuredFields,
             Length: ReqiuredFields.length
         }
@@ -69,6 +71,7 @@ function useEmptyFields({
             console.error("Failed to parse string as JSON in useEmptyFields:", error);
             return {
                 isMising: true,
+                isMissing: true,
                 MissingFields: ReqiuredFields,
                 Length: ReqiuredFields.length
             };
@@ -82,12 +85,14 @@ function useEmptyFields({
     if (Empty_Fields.length > 0) {
         return {
             isMising: true,
+            isMissing: true,
             MissingFields: Empty_Fields,
             Length: Empty_Fields.length
         }
     } else {
         return {
             isMising: false,
+            isMissing: false,
             MissingFields: [],
             Length: 0
         }