+++ src/Types/index.ts
@@ -36,3 +36,4 @@ export * from "./Currency";
 export * from "./Organization";
 export * from "./QR";
 export * from "./Gender";
+export * from "./Controllers";